Consolidated Tokenizer
